+++ src/main/java/isep/web/sakila/webapi/service/CategoryService.java
@@ -6,13 +6,13 @@ import isep.web.sakila.webapi.model.CategoryWO;
 
 public interface CategoryService
 {
-	CategoryWO findById(int id);
+	CategoryWO findById(byte id);
 
 	void saveCategory(CategoryWO categoryWO);
 
 	void updateCategory(CategoryWO categoryWO);
 
-	void deleteCategoryById(int id);
+	void deleteCategoryById(byte id);
 
 	List<CategoryWO> findAllCategories();
